methods trasnfered from BaseView to specific mvp interfaces for HomeWeatherActivity and FindCityActivity so that implementation of interfaces is simplified for every activity
BaseView now only retains bindPresenter and unbindPresenter methods which all Viws shoudl implement

rest of methods which specific Views(Activities) need to implement are in mvp interface of every package

+++ app/src/main/java/com/example/bookee/weatherinfo/findcity/FindCityActivity.java
@@ -66,7 +66,7 @@ public class FindCityActivity extends AppCompatActivity implements mvpView {
         extraData.putString("name",name);
         extraData.putDouble("temp",temperature);
         extraData.putDouble("wind",windSpeed);
-        extraData.putDouble("humid",humidity);
+        extraData.putInt("humid",humidity);
         i.putExtras(extraData);
